add notes form
